+++ NEWS.md
@@ -12,6 +12,8 @@
 
 - 'vignettes/web_only/dimethenamid_2018.rmd': Example evaluations of the dimethenamid data.
 
+- 'intervals': Provide methods of this nlme function for 'nlmixr.mmkin' and 'saem.mmkin' objects.
+
 # mkin 1.0.5 (2021-09-15)
 
 - 'dimethenamid_2018': Correct the data for the Borstel soil. The five observations from Staudenmaier (2013) that were previously stored as "Borstel 2" are actually just a subset of the 16 observations in "Borstel 1" which is now simply "Borstel"